Large Commit: svn merge -r 47809:50286 sync new-upload branch to trunk
git-svn-id: file:///home/git-svn-mirror/pushmi-svn-mirror/branches/new-upload@50288 dd0e9695-b195-4be7-bd10-2dea1a65a6b6

+++ phase3/includes/specials/SpecialBlockip.php
@@ -11,13 +11,11 @@
  */
 function wfSpecialBlockip( $par ) {
 	global $wgUser, $wgOut, $wgRequest;
-
 	# Can't block when the database is locked
 	if( wfReadOnly() ) {
 		$wgOut->readOnlyPage();
 		return;
 	}
-
 	# Permission check
 	if( !$wgUser->isAllowed( 'block' ) ) {
 		$wgOut->permissionRequired( 'block' );
@@ -27,9 +25,9 @@ function wfSpecialBlockip( $par ) {
 	$ipb = new IPBlockForm( $par );
 
 	$action = $wgRequest->getVal( 'action' );
-	if ( 'success' == $action ) {
+	if( 'success' == $action ) {
 		$ipb->showSuccess();
-	} else if ( $wgRequest->wasPosted() && 'submit' == $action &&
+	} else if( $wgRequest->wasPosted() && 'submit' == $action &&
 		$wgUser->matchEditToken( $wgRequest->getVal( 'wpEditToken' ) ) ) {
 		$ipb->doSubmit();
 	} else {
@@ -44,9 +42,10 @@ function wfSpecialBlockip( $par ) {
  */
 class IPBlockForm {
 	var $BlockAddress, $BlockExpiry, $BlockReason;
-#	var $BlockEmail;
+	// The maximum number of edits a user can have and still be hidden
+	const HIDEUSER_CONTRIBLIMIT = 1000;
 
-	function IPBlockForm( $par ) {
+	public function __construct( $par ) {
 		global $wgRequest, $wgUser, $wgBlockAllowsUTEdit;
 
 		$this->BlockAddress = $wgRequest->getVal( 'wpBlockAddress', $wgRequest->getVal( 'ip', $par ) );
@@ -70,7 +69,7 @@ class IPBlockForm {
 		$this->BlockReblock = $wgRequest->getBool( 'wpChangeBlock', false );
 	}
 
-	function showForm( $err ) {
+	public function showForm( $err ) {
 		global $wgOut, $wgUser, $wgSysopUserBans;
 
 		$wgOut->setPagetitle( wfMsg( 'blockip' ) );
@@ -90,17 +89,17 @@ class IPBlockForm {
 		$user = User::newFromName( $this->BlockAddress );
 		
 		$alreadyBlocked = false;
-		if ( $err && $err[0] != 'ipb_already_blocked' ) {
+		if( $err && $err[0] != 'ipb_already_blocked' ) {
 			$key = array_shift($err);
 			$msg = wfMsgReal($key, $err);
 			$wgOut->setSubtitle( wfMsgHtml( 'formerror' ) );
 			$wgOut->addHTML( Xml::tags( 'p', array( 'class' => 'error' ), $msg ) );
-		} elseif ( $this->BlockAddress ) {
+		} elseif( $this->BlockAddress ) {
 			$userId = 0;
-			if ( is_object( $user ) )
+			if( is_object( $user ) )
 				$userId = $user->getId();
 			$currentBlock = Block::newFromDB( $this->BlockAddress, $userId );
-			if ( !is_null($currentBlock) && !$currentBlock->mAuto && # The block exists and isn't an autoblock
+			if( !is_null($currentBlock) && !$currentBlock->mAuto && # The block exists and isn't an autoblock
 				( $currentBlock->mRangeStart == $currentBlock->mRangeEnd || # The block isn't a rangeblock
 				# or if it is, the range is what we're about to block
 				( $currentBlock->mAddress == $this->BlockAddress ) ) ) {
@@ -125,12 +124,11 @@ class IPBlockForm {
 		$scBlockExpiryOptions = wfMsgForContent( 'ipboptions' );
 
 		$showblockoptions = $scBlockExpiryOptions != '-';
-		if (!$showblockoptions)
-			$mIpbother = $mIpbexpiry;
+		if( !$showblockoptions ) $mIpbother = $mIpbexpiry;
 
 		$blockExpiryFormOptions = Xml::option( wfMsg( 'ipbotheroption' ), 'other' );
 		foreach (explode(',', $scBlockExpiryOptions) as $option) {
-			if ( strpos($option, ":") === false ) $option = "$option:$option";
+			if( strpos($option, ":") === false ) $option = "$option:$option";
 			list($show, $value) = explode(":", $option);
 			$show = htmlspecialchars($show);
 			$value = htmlspecialchars($value);
@@ -162,7 +160,7 @@ class IPBlockForm {
 			</tr>
 			<tr>"
 		);
-		if ( $showblockoptions ) {
+		if( $showblockoptions ) {
 			$wgOut->addHTML("
 				<td class='mw-label'>
 					{$mIpbexpiry}
@@ -233,7 +231,7 @@ class IPBlockForm {
 		);
 
 		global $wgSysopEmailBans, $wgBlockAllowsUTEdit;
-		if ( $wgSysopEmailBans && $wgUser->isAllowed( 'blockemail' ) ) {
+		if( $wgSysopEmailBans && $wgUser->isAllowed( 'blockemail' ) ) {
 			$wgOut->addHTML("
 				<tr id='wpEnableEmailBan'>
 					<td>&nbsp;</td>
@@ -247,15 +245,15 @@ class IPBlockForm {
 		}
 
 		// Allow some users to hide name from block log, blocklist and listusers
-		if ( $wgUser->isAllowed( 'hideuser' ) ) {
+		if( $wgUser->isAllowed( 'hideuser' ) ) {
 			$wgOut->addHTML("
 				<tr id='wpEnableHideUser'>
 					<td>&nbsp;</td>
-					<td class='mw-input'>" .
+					<td class='mw-input'><strong>" .
 						Xml::checkLabel( wfMsg( 'ipbhidename' ),
 							'wpHideName', 'wpHideName', $this->BlockHideName,
 							array( 'tabindex' => '10' ) ) . "
-					</td>
+					</strong></td>
 				</tr>"
 			);
 		}
@@ -328,12 +326,12 @@ class IPBlockForm {
 		$rxIP = "($rxIP4|$rxIP6)";
 
 		# Check for invalid specifications
-		if ( !preg_match( "/^$rxIP$/", $this->BlockAddress ) ) {
+		if( !preg_match( "/^$rxIP$/", $this->BlockAddress ) ) {
 			$matches = array();
-		  	if ( preg_match( "/^($rxIP4)\\/(\\d{1,2})$/", $this->BlockAddress, $matches ) ) {
+		  	if( preg_match( "/^($rxIP4)\\/(\\d{1,2})$/", $this->BlockAddress, $matches ) ) {
 		  		# IPv4
-				if ( $wgSysopRangeBans ) {
-					if ( !IP::isIPv4( $this->BlockAddress ) || $matches[2] < 16 || $matches[2] > 32 ) {
+				if( $wgSysopRangeBans ) {
+					if( !IP::isIPv4( $this->BlockAddress ) || $matches[2] < 16 || $matches[2] > 32 ) {
 						return array('ip_range_invalid');
 					}
 					$this->BlockAddress = Block::normaliseRange( $this->BlockAddress );
@@ -341,10 +339,10 @@ class IPBlockForm {
 					# Range block illegal
 					return array('range_block_disabled');
 				}
-			} else if ( preg_match( "/^($rxIP6)\\/(\\d{1,3})$/", $this->BlockAddress, $matches ) ) {
+			} else if( preg_match( "/^($rxIP6)\\/(\\d{1,3})$/", $this->BlockAddress, $matches ) ) {
 		  		# IPv6
-				if ( $wgSysopRangeBans ) {
-					if ( !IP::isIPv6( $this->BlockAddress ) || $matches[2] < 64 || $matches[2] > 128 ) {
+				if( $wgSysopRangeBans ) {
+					if( !IP::isIPv6( $this->BlockAddress ) || $matches[2] < 64 || $matches[2] > 128 ) {
 						return array('ip_range_invalid');
 					}
 					$this->BlockAddress = Block::normaliseRange( $this->BlockAddress );
@@ -354,7 +352,7 @@ class IPBlockForm {
 				}
 			} else {
 				# Username block
-				if ( $wgSysopUserBans ) {
+				if( $wgSysopUserBans ) {
 					$user = User::newFromName( $this->BlockAddress );
 					if( !is_null( $user ) && $user->getId() ) {
 						# Use canonical name
@@ -369,15 +367,15 @@ class IPBlockForm {
 			}
 		}
 
-		if ( $wgUser->isBlocked() && ( $wgUser->getId() !== $userId ) ) {
+		if( $wgUser->isBlocked() && ( $wgUser->getId() !== $userId ) ) {
 			return array( 'cant-block-while-blocked' );
 		}
 
 		$reasonstr = $this->BlockReasonList;
-		if ( $reasonstr != 'other' && $this->BlockReason != '' ) {
+		if( $reasonstr != 'other' && $this->BlockReason != '' ) {
 			// Entry from drop down menu + additional comment
 			$reasonstr .= wfMsgForContent( 'colon-separator' ) . $this->BlockReason;
-		} elseif ( $reasonstr == 'other' ) {
+		} elseif( $reasonstr == 'other' ) {
 			$reasonstr = $this->BlockReason;
 		}
 
@@ -385,21 +383,30 @@ class IPBlockForm {
 		if( $expirestr == 'other' )
 			$expirestr = $this->BlockOther;
 
-		if ( ( strlen( $expirestr ) == 0) || ( strlen( $expirestr ) > 50) ) {
+		if( ( strlen( $expirestr ) == 0) || ( strlen( $expirestr ) > 50) ) {
 			return array('ipb_expiry_invalid');
 		}
 		
-		if ( false === ($expiry = Block::parseExpiryInput( $expirestr )) ) {
+		if( false === ($expiry = Block::parseExpiryInput( $expirestr )) ) {
 			// Bad expiry.
 			return array('ipb_expiry_invalid');
 		}
 		
-		if( $this->BlockHideName && $expiry != 'infinity' ) {
-			// Bad expiry.
-			return array('ipb_expiry_temp');
+		if( $this->BlockHideName ) {
+			// Recheck params here...
+			if( !$userId || !$wgUser->isAllowed('hideuser') ) {
+				$this->BlockHideName = false; // IP users should not be hidden
+			} else if( $expiry !== 'infinity' ) {
+				// Bad expiry.
+				return array('ipb_expiry_temp');
+			} else if( User::edits($userId) > self::HIDEUSER_CONTRIBLIMIT ) {
+				// Typically, the user should have a handful of edits.
+				// Disallow hiding users with many edits for performance.
+				return array('ipb_hide_invalid');
+			}
 		}
 
-		# Create block
+		# Create block object
 		# Note: for a user block, ipb_address is only for display purposes
 		$block = new Block( $this->BlockAddress, $userId, $wgUser->getId(),
 			$reasonstr, wfTimestampNow(), 0, $expiry, $this->BlockAnonOnly,
@@ -407,11 +414,15 @@ class IPBlockForm {
 			$this->BlockEmail, isset( $this->BlockAllowUsertalk ) ? $this->BlockAllowUsertalk : $wgBlockAllowsUTEdit
 		);
 
-		if ( wfRunHooks('BlockIp', array(&$block, &$wgUser)) ) {
-
-			if ( !$block->insert() ) {
-				if ( !$this->BlockReblock ) {
+		# Should this be privately logged?
+		$suppressLog = (bool)$this->BlockHideName;
+		if( wfRunHooks('BlockIp', array(&$block, &$wgUser)) ) {
+			# Try to insert block. Is there a conflicting block?
+			if( !$block->insert() ) {
+				# Show form unless the user is already aware of this...
+				if( !$this->BlockReblock ) {
 					return array( 'ipb_already_blocked' );
+				# Otherwise, try to update the block...
 				} else {
 					# This returns direct blocks before autoblocks/rangeblocks, since we should
 					# be sure the user is blocked by now it should work for our purposes
@@ -419,19 +430,34 @@ class IPBlockForm {
 					if( $block->equals( $currentBlock ) ) {
 						return array( 'ipb_already_blocked' );
 					}
+					# If the name was hidden and the blocking user cannot hide
+					# names, then don't allow any block changes...
+					if( $currentBlock->mHideName && !$wgUser->isAllowed('hideuser') ) {
+						return array( 'hookaborted' );
+					}
 					$currentBlock->delete();
 					$block->insert();
+					# If hiding/unhiding a name, this should go in the private logs
+					$suppressLog = $suppressLog || (bool)$currentBlock->mHideName;
 					$log_action = 'reblock';
+					# Unset _deleted fields if requested
+					if( $currentBlock->mHideName && !$this->BlockHideName ) {
+						self::unsuppressUserName( $this->BlockAddress, $userId );
+					}
 				}
 			} else {
 				$log_action = 'block';
 			}
 			wfRunHooks('BlockIpComplete', array($block, $wgUser));
 
-			if ( $this->BlockWatchUser &&
-				# Only show watch link when this is no range block
-				$block->mRangeStart == $block->mRangeEnd) {
-				$wgUser->addWatch ( Title::makeTitle( NS_USER, $this->BlockAddress ) );
+			# Set *_deleted fields if requested
+			if( $this->BlockHideName ) {
+				self::suppressUserName( $this->BlockAddress, $userId, $reasonstr );
+			}
+
+			# Only show watch link when this is no range block
+			if( $this->BlockWatchUser && $block->mRangeStart == $block->mRangeEnd ) {
+				$wgUser->addWatch( Title::makeTitle( NS_USER, $this->BlockAddress ) );
 			}
 			
 			# Block constructor sanitizes certain block options on insert
@@ -444,27 +470,94 @@ class IPBlockForm {
 			$logParams[] = $this->blockLogFlags();
 
 			# Make log entry, if the name is hidden, put it in the oversight log
-			$log_type = ($this->BlockHideName) ? 'suppress' : 'block';
+			$log_type = $suppressLog ? 'suppress' : 'block';
 			$log = new LogPage( $log_type );
 			$log->addEntry( $log_action, Title::makeTitle( NS_USER, $this->BlockAddress ),
-			  $reasonstr, $logParams );
+				$reasonstr, $logParams );
 
 			# Report to the user
 			return array();
-		}
-		else
+		} else {
 			return array('hookaborted');
+		}
+	}
+	
+	public static function suppressUserName( $name, $userId, $reason = '' ) {
+		$user = User::newFromName( $name, false );
+		# Delete the user pages that exists
+		$title = $user->getUserPage();
+		if( ($id = $title->getArticleID(GAID_FOR_UPDATE)) ) {
+			$article = new Article( $title );
+			$article->doDeleteArticle( $reason, true /*suppress*/, $id );
+		}
+		# Delete the user talk pages that exists
+		$title = $user->getTalkPage();
+		if( $id = $title->getArticleID(GAID_FOR_UPDATE) ) {
+			$article = new Article( $title );
+			$article->doDeleteArticle( $reason, true /*suppress*/, $id );
+		}
+		$op = '|'; // bitwise OR
+		return self::setUsernameBitfields( $name, $userId, $op );
+	}
+	
+	public static function unsuppressUserName( $name, $userId ) {
+		$op = '&'; // bitwise AND
+		return self::setUsernameBitfields( $name, $userId, $op );
+	}
+	
+	private static function setUsernameBitfields( $name, $userId, $op ) {
+		if( $op !== '|' && $op !== '&' ) return false; // sanity check
+		$dbw = wfGetDB( DB_MASTER );
+		$delUser = Revision::DELETED_USER | Revision::DELETED_RESTRICTED;
+		$delAction = LogPage::DELETED_ACTION | Revision::DELETED_RESTRICTED;
+		# Normalize user name
+		$userTitle = Title::makeTitleSafe( NS_USER, $name );
+		$userDbKey = $userTitle->getDBKey();
+		# To suppress, we OR the current bitfields with Revision::DELETED_USER
+		# to put a 1 in the username *_deleted bit. To unsuppress we AND the
+		# current bitfields with the inverse of Revision::DELETED_USER. The
+		# username bit is made to 0 (x & 0 = 0), while others are unchanged (x & 1 = x).
+		# The same goes for the sysop-restricted *_deleted bit.
+		if( $op == '&' ) {
+			$delUser = "~{$delUser}";
+			$delAction = "~{$delAction}";
+		}
+		# Hide name from live edits
+		$dbw->update( 'revision', array("rev_deleted = rev_deleted $op $delUser"),
+			array('rev_user' => $userId), __METHOD__ );
+		# Hide name from deleted edits
+		$dbw->update( 'archive', array("ar_deleted = ar_deleted $op $delUser"),
+			array('ar_user_text' => $name), __METHOD__ );
+		# Hide name from logs
+		$dbw->update( 'logging', array("log_deleted = log_deleted $op $delUser"),
+			array('log_user' => $userId, "log_type != 'suppress'"), __METHOD__ );
+		$dbw->update( 'logging', array("log_deleted = log_deleted $op $delAction"),
+			array('log_namespace' => NS_USER, 'log_title' => $userDbKey,
+				"log_type != 'suppress'"), __METHOD__ );
+		# Hide name from RC
+		$dbw->update( 'recentchanges', array("rc_deleted = rc_deleted $op $delUser"),
+			array('rc_user_text' => $name), __METHOD__ );
+		$dbw->update( 'recentchanges', array("rc_deleted = rc_deleted $op $delAction"),
+			array('rc_namespace' => NS_USER, 'rc_title' => $userDbKey, 'rc_logid > 0'), __METHOD__ );
+		# Hide name from live images
+		$dbw->update( 'oldimage', array("oi_deleted = oi_deleted $op $delUser"),
+			array('oi_user_text' => $name), __METHOD__ );
+		# Hide name from deleted images
+		# WMF - schema change pending
+		# $dbw->update( 'filearchive', array("fa_deleted = fa_deleted $op $delUser"),
+		#	array('fa_user_text' => $name), __METHOD__ );
+		# Done!
+		return true;
 	}
 
 	/**
 	 * UI entry point for blocking
 	 * Wraps around doBlock()
 	 */
-	function doSubmit()
-	{
+	public function doSubmit() {
 		global $wgOut;
 		$retval = $this->doBlock();
-		if(empty($retval)) {
+		if( empty($retval) ) {
 			$titleObj = SpecialPage::getTitleFor( 'Blockip' );
 			$wgOut->redirect( $titleObj->getFullURL( 'action=success&ip=' .
 				urlencode( $this->BlockAddress ) ) );
@@ -473,7 +566,7 @@ class IPBlockForm {
 		$this->showForm( $retval );
 	}
 
-	function showSuccess() {
+	public function showSuccess() {
 		global $wgOut;
 
 		$wgOut->setPagetitle( wfMsg( 'blockip' ) );
@@ -482,11 +575,11 @@ class IPBlockForm {
 		$wgOut->addHTML( $text );
 	}
 
-	function showLogFragment( $out, $title ) {
+	private function showLogFragment( $out, $title ) {
 		global $wgUser;
 		$out->addHTML( Xml::element( 'h2', NULL, LogPage::logName( 'block' ) ) );
 		$count = LogEventsList::showLogExtract( $out, 'block', $title->getPrefixedText(), '', 10 );
-		if($count > 10){
+		if( $count > 10 ) {
 			$out->addHTML( $wgUser->getSkin()->link(
 				SpecialPage::getTitleFor( 'Log' ),
 				wfMsgHtml( 'blocklog-fulllog' ),
@@ -494,6 +587,12 @@ class IPBlockForm {
 				array(
 					'type' => 'block',
 					'page' => $title->getPrefixedText() ) ) );
+		}
+		// Add suppression block entries if allowed
+		if( $wgUser->isAllowed('hideuser') ) {
+			$out->addHTML( Xml::element( 'h2', NULL, LogPage::logName( 'suppress' ) ) );
+			LogEventsList::showLogExtract( $out, 'suppress', $title->getPrefixedText(), '',
+				10, array('log_action' => array('block','reblock','unblock')) );
 		}
 	}
 
@@ -511,12 +610,15 @@ class IPBlockForm {
 			$flags[] = 'anononly';
 		if( $this->BlockCreateAccount )
 			$flags[] = 'nocreate';
-		if( !$this->BlockEnableAutoblock )
+		if( !$this->BlockEnableAutoblock && !IP::isIPAddress( $this->BlockAddress ) )
+			// Same as anononly, this is not displayed when blocking an IP address
 			$flags[] = 'noautoblock';
-		if ( $this->BlockEmail )
+		if( $this->BlockEmail )
 			$flags[] = 'noemail';
-		if ( !$this->BlockAllowUsertalk && $wgBlockAllowsUTEdit )
+		if( !$this->BlockAllowUsertalk && $wgBlockAllowsUTEdit )
 			$flags[] = 'nousertalk';
+		if( $this->BlockHideName )
+			$flags[] = 'hiddenname';
 		return implode( ',', $flags );
 	}
